refactor: httpmethod 패키지 request 아래로 이동
