+++ src/main/java/webserver/request/method/exception/InvalidHttpMethodException.java
@@ -1,4 +1,4 @@
-package webserver.method.exception;
+package webserver.request.method.exception;
 
 public class InvalidHttpMethodException extends RuntimeException{
     public InvalidHttpMethodException(String message) {
