+++ src/test/java/webserver/request/header/HttpMethodTest.java
@@ -6,8 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import webserver.method.HttpMethod;
-import webserver.method.exception.InvalidHttpMethodException;
+import webserver.request.method.HttpMethod;
+import webserver.request.method.exception.InvalidHttpMethodException;
 
 class HttpMethodTest {
